+++ projects/gateway2/deployer/deployer_test.go
@@ -1269,7 +1269,7 @@ var _ = Describe("Deployer", func() {
 								Ports: []*gw2_v1alpha1.Port{
 									{
 										Port:     80,
-										NodePort: ptr.To(int32(30000)),
+										NodePort: ptr.To(uint16(30000)),
 									},
 								},
 							},
